% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_characteristic.population}
\alias{add_characteristic.population}
\title{Add characteristic to a population}
\usage{
\method{add_characteristic}{population}(x, name, value = NA)
}
\arguments{
\item{x}{Object of \link{population} class representing a population.}

\item{name}{Name of the characteristic to add.}

\item{value}{Value of the characteristic. By default \code{NA}.}
}
\description{
Add characteristic to a population
}
