% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_characteristic}
\alias{add_characteristic}
\title{Generic method for add_characteristic}
\usage{
add_characteristic(x, name, value = NA)
}
\arguments{
\item{x}{An object.}

\item{name}{Name of the characteristic to add.}

\item{value}{Value of the characteristic. By default \code{NA}.}
}
\description{
Generic method for add_characteristic
}
