% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{age_pyramids.population}
\alias{age_pyramids.population}
\title{Age pyramid from a population data frame at some given times.}
\usage{
\method{age_pyramids}{population}(object, time = 0, ages = c(0:110, Inf))
}
\arguments{
\item{object}{Object of \code{\link{population}} class representing a population.}

\item{time}{The age pyramid is computed at instant \code{time}. Must be a numeric greater than or equal to 0.}

\item{ages}{\emph{(Optional)} A numeric vector of distinct positive values composing age groups. Must be in increasing order.}
}
\description{
Vectorial version in time of the function \code{\link{age_pyramid.population}}. Not compatible with IBMs including swap events.
}
\details{
For convenience. This is a just a \code{lapply} call of \code{age_pyramid.population} on the vector \code{time}.
}
