% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{check_intensity_code}
\alias{check_intensity_code}
\title{Check the intensity code.}
\usage{
check_intensity_code(code)
}
\arguments{
\item{code}{String containing the intensity code.}
}
\description{
Verifies that the intensity contains the string 'result'.
}
