% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{check_interaction_code}
\alias{check_interaction_code}
\title{Check the interaction code.}
\usage{
check_interaction_code(code)
}
\arguments{
\item{code}{String containing the interaction code.}
}
\description{
Verifies that the interaction contains the string 'result'.
}
