% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{check_kernel_code}
\alias{check_kernel_code}
\title{Check the kernel code.}
\usage{
check_kernel_code(code)
}
\arguments{
\item{code}{String containing the kernel code.}
}
\description{
Verifies the kernel code.
}
