% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compatibility_chars_events}
\alias{compatibility_chars_events}
\title{Check characteristics-events compatibility}
\usage{
compatibility_chars_events(characteristics, events)
}
\arguments{
\item{characteristics}{List of characteristics}

\item{events}{List of events}
}
\description{
A function to check the compatibility between characteristics and events
}
