% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compatibility_pop_model}
\alias{compatibility_pop_model}
\title{Check population-model compatibility}
\usage{
compatibility_pop_model(pop, model)
}
\arguments{
\item{pop}{An object of class \link{population}}

\item{model}{An Individual Based Model created with the \link{mk_model} function}
}
\description{
A function to check the compatibility between a population and a model
}
