% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{details_intensity_code}
\alias{details_intensity_code}
\title{Autogenerated documentation for intensity code}
\usage{
details_intensity_code()
}
\description{
Autogenerated documentation for intensity code
}
\keyword{internal}
