% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{details_interaction_code}
\alias{details_interaction_code}
\title{Autogenerated documentation for interaction code}
\usage{
details_interaction_code()
}
\description{
Autogenerated documentation for interaction code
}
\keyword{internal}
