% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{details_kernel_code}
\alias{details_kernel_code}
\title{Autogenerated documentation for kernel code}
\usage{
details_kernel_code()
}
\description{
Autogenerated documentation for kernel code
}
\keyword{internal}
