% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{details_type_event}
\alias{details_type_event}
\title{Autogenerated documentation for event type}
\usage{
details_type_event()
}
\description{
Autogenerated documentation for event type
}
\keyword{internal}
