% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EW_pop_14}
\alias{EW_pop_14}
\title{England and Wales (EW) 2014 population, death and birth rates.}
\format{
A list containing:
\describe{
\item{\code{age_pyramid}}{Data frame containing EW age pyramid for year 2014, by gender and single year of age (0-115).}
\item{\code{rates}}{A list containing three data frames:\describe{
\item{\code{birth}}{Birth rates data frame, by age of mother and 5 years age groups.}
\item{\code{death_male}}{Male death rates data frame, by single year of age (0-90+).}
\item{\code{death_female}}{Female death rates dataframe, by single year of age (0-90+).}}}
\item{\code{sample}}{Population dataframe composed of 100 000 individuals, sampled from \code{age_pyramid}.}
}
}
\usage{
EW_pop_14
}
\description{
EW 2014 population and death rates by age  and gender  (Source: Office for National Statistics, reference number 006518).

Female birth rates by age of the mother (Source: Office for National Statistics birth summary tables).
}
\keyword{datasets}
