% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EW_pop_out}
\alias{EW_pop_out}
\title{Example of "human population" after 100 years of simulation.}
\format{
Data frame containing a population structured by age and gender,
simulated with an initial population of 100 000 individuals sampled from \code{EW_pop_14$age_pyramid}
over 100 years, with birth and death events.
}
\usage{
EW_pop_out
}
\description{
Example of "human population"  data frame after 100 years of simulation, based on a sample of England and Wales 2014 population and demographic rates.
}
\keyword{datasets}
