% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EW_popIMD_14}
\alias{EW_popIMD_14}
\title{England and Wales (EW) 2014 population and death rates by Index of Multiple Deprivation (IMD).}
\format{
A list containing:
\describe{
\item{\code{age_pyramid}}{Data frame containing EW age pyramid for year 2014, by gender, IMD and single year of age (0-115).

Individuals in the age class 90+ are distributed in the single year of age classes as in the EW population.}
\item{\code{death_rates}}{List containing 4 fields:
\describe{
\item{\code{male}}{ Male death rates data frame, by IMD and single year of age (0-90+).}
\item{\code{female}}{ Female death rates dataframe, by IMD and single year of age (0-90+).}
}
}
\item{\code{sample}}{Population dataframe composed of 100 000 individuals, sampled from \code{age_pyramid}.}
}
}
\usage{
EW_popIMD_14
}
\description{
EW population, death rates by age, gender and IMD for year 2014 (Source: Office for National Statistics, reference number 006518).
}
\keyword{datasets}
