% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EWdata_hmd}
\alias{EWdata_hmd}
\title{England and Wale mortality data (source: Human Mortality Database)}
\format{
An object of class \code{demogdata} of length 7.
}
\usage{
EWdata_hmd
}
\description{
Obtained with

\code{EWdata_hmd <- hmd.mx(country = "GBRTENW", username = ... , password = ...,label = "England and Wales")}
}
\keyword{datasets}
