% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{exposure_table}
\alias{exposure_table}
\title{Creates an exposure table.}
\usage{
exposure_table(population, ages, period)
}
\arguments{
\item{population}{Population data frame containing at least \code{'birth'} and \code{'death'} columns.}

\item{ages}{A vector of size \code{N} composed of age groups.
The function computes the number of death  in each age group \code{[ages[i],ages[i+1])}, for \code{i=1..N-1}.}

\item{period}{A vector of size \code{M} composed of time intervals.}
}
\value{
An exposure matrix
}
\description{
Returns the Central Exposure-to-Risk for given ages groups and time period.
The central Exposure-to-risk is computed as the sum of the time spent by individuals in a given age group over a given period, where age is the age at last birthday.
}
\details{
The function computes the central exposure-to-risk in each time interval \code{[t[j],t[j+1])}, \code{j=1..M}, and age groups.
}
\examples{
ex_table <- exposure_table(EW_pop_out,0:101,0:2)

}
