% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_characteristics.population}
\alias{get_characteristics.population}
\title{Returns names and C types of the characteristics.}
\usage{
\method{get_characteristics}{population}(object, ...)
}
\arguments{
\item{object}{Object of \link{population} class representing a population.}

\item{...}{additional arguments.}
}
\value{
Named vector composed of characteristics names and C types. If the population has no characteristics, which means that it has only the \code{birth} and \code{death} columns, this returns \code{NULL}.
}
\description{
Returns names and C types of the characteristics (other than birth and death) of the individuals in a population, from a population data frame.
}
\examples{
get_characteristics(population(EW_pop_14$sample))

}
