% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{has_event_type}
\alias{has_event_type}
\title{Search for a given string in event types}
\usage{
has_event_type(events, str)
}
\arguments{
\item{events}{list of events}

\item{str}{string to search}
}
\value{
boolean
}
\description{
Search for a given string in event types
}
\keyword{internal}
