% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim.R
\name{id_complete}
\alias{id_complete}
\title{Complete a population id}
\usage{
id_complete(population)
}
\arguments{
\item{population}{population dataframe}
}
\description{
In a population with an \code{id} column, verifies that there are no NA \code{id} values,
and if there are, completes the missing \code{id}s
}
\keyword{internal}
