% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linfun}
\alias{linfun}
\title{Linear interpolation function.}
\usage{
linfun(x, y, yleft = y[1], yright = y[length(y)])
}
\arguments{
\item{x, y}{Numeric vectors giving the coordinates of the points to be interpolated.}

\item{yleft}{The value to be returned when input \code{x} values are less than \code{min(x)}.}

\item{yright}{The value to be returned when input \code{x} values are greater than \code{max(x)}.}
}
\value{
Objet of class \code{linfun} and \code{function} which is an \code{\link[stats]{approxfun}} function with \code{method = 'linear'}.
}
\description{
Return a function performing the linear interpolation.
}
\details{
A C++ version of this function is available. See \code{vignette('IBMPopSim_cpp')} for more details.
}
