% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{max.stepfun}
\alias{max.stepfun}
\title{Returns the maximum of a function of class \code{stepfun}.}
\usage{
\method{max}{stepfun}(..., na.rm = FALSE)
}
\arguments{
\item{...}{argument of class \code{stepfun}}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\value{
The maximum of the step function.
}
\description{
Returns the maximum of a function of class \code{stepfun}.
}
