% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_pop_withid}
\alias{merge_pop_withid}
\title{A function returning a merged dataframe from a list of population dataframes with id.}
\usage{
merge_pop_withid(pop_df_list, chars_tracked = NULL)
}
\arguments{
\item{pop_df_list}{A list of population dataframe where the first three columns of each dataframe are id, birth and death.}

\item{chars_tracked}{A vector of characteristics to be tracked over time.}
}
\value{
A dataframe composed of all individuals with their characteristics at each simulation time.
}
\description{
A function returning a merged dataframe from a list of population dataframes with id.
}
