% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mk_parameters_types}
\alias{mk_parameters_types}
\title{Extract types (R and C++) of the parameters}
\usage{
mk_parameters_types(parameters)
}
\arguments{
\item{parameters}{parameters}
}
\value{
list
}
\description{
Extract types (R and C++) of the parameters
}
\keyword{internal}
