% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcpp_from_template.R
\name{mkcpp_declaration_individual}
\alias{mkcpp_declaration_individual}
\title{Function for internal use}
\usage{
mkcpp_declaration_individual(ind_t)
}
\description{
Build source code for individual creation.
}
\keyword{internal}
