% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcpp_from_template.R
\name{mkcpp_definition_events}
\alias{mkcpp_definition_events}
\title{Function for internal use}
\usage{
mkcpp_definition_events(events)
}
\description{
Build source code for event definition.
}
\keyword{internal}
