% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcpp_from_template.R
\name{mkcpp_initialisation_population}
\alias{mkcpp_initialisation_population}
\title{Function for internal use}
\usage{
mkcpp_initialisation_population(ind_t)
}
\description{
Build source code for population initialization.
}
\keyword{internal}
