% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcpp_from_template.R
\name{mkcpp_parameters}
\alias{mkcpp_parameters}
\title{Function for internal use}
\usage{
mkcpp_parameters(parameters, parameters_t)
}
\description{
Build source code for parameters creation.
}
\keyword{internal}
