% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{population_alive}
\alias{population_alive}
\title{Generic method for population_alive}
\usage{
population_alive(object, t, a1 = 0, a2 = Inf, ...)
}
\arguments{
\item{object}{A population.}

\item{t}{A numeric indicating the time at which alive individuals are observed.}

\item{a1}{0 by default. Lower bound for age.}

\item{a2}{Inf by default. Upper bound for age.}

\item{...}{Additional params.}
}
\value{
All individuals alive at time \code{t} and of age in \code{[a1,a2)}.
}
\description{
Generic method for population_alive
}
