% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{print.event}
\alias{print.event}
\title{Print Event}
\usage{
\method{print}{event}(x, ...)
}
\arguments{
\item{x}{Argument of class \code{event}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
\code{print} method for class "event" giving a short description of an event.
}
