% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.model}
\alias{print.model}
\title{Printing of a model}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{argument of class \code{model}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{print} method for class \code{model}.
}
