% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.population}
\alias{print.population}
\title{Printing population}
\usage{
\method{print}{population}(x, ...)
}
\arguments{
\item{x}{Object of \link{population} class representing a population.}

\item{...}{Additional arguments}
}
\value{
Print the population
}
\description{
Print a population
}
