% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\docType{class}
\name{pyramid}
\alias{pyramid}
\title{Class pyramid}
\usage{
pyramid(x)
}
\arguments{
\item{x}{Data frame, containing at least an \code{age} and a \code{value} column}
}
\value{
Given data frame augmented of the "age_pyramid" class.
}
\description{
Data frame containing an age pyramid, with at least an \code{age} and a \code{value} column,
and eventually some other characteristics. If a \code{male} column is present, it must be a logical vector,
if a \code{group} column is present, it must be a vector of type character.
}
