% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{summary.event}
\alias{summary.event}
\title{Summarizing an event}
\usage{
\method{summary}{event}(object, ...)
}
\arguments{
\item{object}{Argument of class \code{event}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
\code{summary} method for class \code{event} giving a detailed description of an event.
}
