% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.logs}
\alias{summary.logs}
\title{Summary logs}
\usage{
\method{summary}{logs}(object, ...)
}
\arguments{
\item{object}{Logs of the output of a call to \link{popsim} function}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
Print column names and number of individuals
}
\description{
Summary of the logs of a simulation
}
