% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.model}
\alias{summary.model}
\title{Summary of a model}
\usage{
\method{summary}{model}(object, ...)
}
\arguments{
\item{object}{argument of class \code{model}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{summary} method for class \code{model}.
}
