% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.population}
\alias{summary.population}
\title{Summary population}
\usage{
\method{summary}{population}(object, ...)
}
\arguments{
\item{object}{Object of \link{population} class representing a population.}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
Print column names and number of individuals
}
\description{
Summary of a population with column names and number of individuals
}
