% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.simulation_output}
\alias{summary.simulation_output}
\title{Summary simulation output}
\usage{
\method{summary}{simulation_output}(object, ...)
}
\arguments{
\item{object}{Output of a call to \link{popsim} function}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
Summary of population(s) and the logs
}
\description{
Summary of a simulation output
}
