% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_params}
\alias{toy_params}
\title{Toy parameters for IBMPopSim-human_popIMD vignette example.}
\format{
A list containing:
\describe{
\item{\code{birth}}{A list of 3 numeric vectors  (\code{alpha}, \code{beta}, \code{TFR_weights}) for creating birth intensity with the Weibull probability density function.}
\item{\code{swap}}{A List of one numeric vector and two data frames  (\code{ages}, \code{intensities} and \code{distribution}) for creating the swap intensity and kernel functions.}
}
}
\usage{
toy_params
}
\description{
Toy parameters for IBMPopSim-human_popIMD vignette example.
}
\keyword{datasets}
