% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{age_pyramid}
\alias{age_pyramid}
\title{Generic method for age_pyramid}
\usage{
age_pyramid(object, time = 0, ages = c(0:110, Inf), ...)
}
\arguments{
\item{object}{Population.}

\item{time}{The age pyramid is computed at instant \code{time}. Must be a numeric greater than or equal to 0.}

\item{ages}{\emph{(Optional)} A numeric vector of distinct positive values composing age groups. Must be in increasing order.}

\item{...}{Additional parameters}
}
\value{
An object of class \code{\link{pyramid}} containing the age pyramid of a population at instant \code{time}.
}
\description{
Generic method for age_pyramid
}
