% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{age_pyramids}
\alias{age_pyramids}
\title{Generic method for age_pyramids}
\usage{
age_pyramids(object, time = 0, ages = c(0:110, Inf))
}
\arguments{
\item{object}{Population.}

\item{time}{The age pyramid is computed at instant \code{time}. Must be a numeric greater than or equal to 0.}

\item{ages}{\emph{(Optional)} A numeric vector of distinct positive values composing age groups. Must be in increasing order.}
}
\description{
Generic method for age_pyramids
}
