% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_characteristics}
\alias{get_characteristics}
\title{Generic method for get_characteristics}
\usage{
get_characteristics(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional parameters.}
}
\description{
Generic method for get_characteristics
}
