% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim.R
\name{id_complete}
\alias{id_complete}
\title{Complete a population id}
\usage{
id_complete(object, ...)
}
\arguments{
\item{object}{A data frame containing at least a column \code{birth} and \code{death}, that will be converted in \link{population} class.}
}
\description{
In a population with an \code{id} column, verifies that there are no NA \code{id} values,
and if there are, completes the missing \code{id}s
}
\keyword{internal}
