% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{merge_pop_withid}
\alias{merge_pop_withid}
\title{A function returning a merged dataframe from a list of population dataframes with id.}
\usage{
merge_pop_withid(pop_df_list, char_ignored = NULL)
}
\arguments{
\item{pop_df_list}{A list of population dataframe where the first three columns of each dataframe are id, birth and death.}

\item{char_ignored}{A vector of characteristics names which are only kept from the first element of pop_df_list.}
}
\value{
A dataframe composed of all individuals with their characteristics at each simulation time.
}
\description{
A function returning a merged dataframe from a list of population dataframes with id.
}
