% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{popsample}
\alias{popsample}
\title{Generic method for popsample}
\usage{
popsample(age_pyramid, size, age_max = 120, time = 0)
}
\arguments{
\item{age_pyramid}{Age pyramid.}

\item{size}{A non-negative integer giving the number of individuals in population.}

\item{age_max}{\emph{(Optional)} A non-negative numeric which replace (if exists) the \code{Inf} in \code{\link{age_pyramid.population}}.}

\item{time}{\emph{(Optional)} The age pyramid is computed at instant \code{time}. Must be a numeric greater than or equal to 0.}
}
\value{
Object of \code{\link{population}} class representing a data frame of size \code{size} containing a population of individuals.
}
\description{
Generic method for popsample
}
