% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{population_alive}
\alias{population_alive}
\title{Returns a population of individuals alive.}
\usage{
population_alive(population, t, a1 = 0, a2 = Inf)
}
\arguments{
\item{population}{A population data frame containing at least a column \code{'birth'} and \code{'death'}.}

\item{t}{A numeric indicating the time.}

\item{a1}{0 by default. All individuals of age over \code{a1} at \code{t} are selected.}

\item{a2}{Inf by default. All individuals of age below \code{a2} at \code{t} are selected.}
}
\value{
The function returns a population data frame containing all individuals alive at time \code{t} and of age in \code{[a1,a2)}.
}
\description{
Returns a population of individuals alive.
}
