% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{summary.event}
\alias{summary.event}
\title{Summary of an event.}
\usage{
\method{summary}{event}(object, ...)
}
\arguments{
\item{object}{Argument of class \code{event}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Summary of an event.
}
